backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pymysql==1.1.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
